Add input reading
Adds the ability to read in the input. Done by first reading into a matrix which is then converted to dictionary.

Uses the ast.literal_eval function in order to get proper datatypes. However this function chokes on strings that contain numbers
therefore a try, except statement is used in order to catch these errors and convert the offending inputs into strings.