Resolves issue #6
The error was due to an incorrect catalog being passed to textdump. This was solved by passing it manually.

The source photometry is fully implemented. Most major change is the adding of a ```try: ... except: ...``` statement to the filter selection.
This was done because the 814 has the filter under the FILTER2 key instead of the FILTER key