Add ACS zeropoint function
Adds a function for calculating acs zeropoints from the info in
the header. This is then used in the try except statement in phot()
so it is used if no filtermatch is found in the list.

Also fixes the catching of the mean of empty slice warning in the
apcorr file and gives a custom warning instead. This also sets the
apcorr and apcorr error to 0 i.e. no apcorr is applied.