Fix breaking on F140LP filter
When the pipeline was run on the F140LP filter, it broke since it was
unable to detect the filter in the header. This fixes this by looking
at both the FILTER, FILTER1 and FILTER2 keywords.

A further crash was caused by defining the labels to be 4 characters
long when parsing the filter form the photometric catalog names.
This was fixed by instead allowing it to select the full filter name
using a split('.') to separate out the file ending.

Resolves: #25